Add beginnings of a calculation architecture
This architecture is meant to separate the processing of form data from
the logic of iterating through the candidates.

Each calculator class is instantiated with a list of candidates. It is
that class's responsibility to (inside a "fetch" method) do whatever
calculations it needs based on the raw data in the `efile_COAK_2016_*`
tables and save the results with:

  candidate.save_calculation([name], [value])

which is idempotent. Then at API serialization time, we fetch that
calculation with

  candidate.calculation(name)

and serialize it in the response.