Autodiscover / run calculators that are named correctly
This adds a bit of ruby magic to the `process.rb` script by converting
calculator filenames into class names and attempting to use them at
runtime.

For example, the file

  calculators/foo_bar_calculator.rb

would be expected to define the class named:

  FooBarCalculator

This gives us the advantage that we can auto-load the calculators and
run them without having to be explicitly listed inside of process.rb.

I also added a bunch of documentation about how to write a calculator.