Add ballot measure supporters / opponents
This adds a simple calculation based on Schedule E to determine a ballot
measure's supporters and opponents. The matching is a bit fuzzy - as of
writing the "Bal_Num" column contains only "TBD", so we can't easily
match up the expenditures to the ballot measure except by hardcoding a
mapping of the four "Bal_Name" values.

This branch doesn't support using Filer IDs or a custom ID for these
committees because there would be nowhere for those links to go, and I
want to punt on figuring out how to model the committees until later.

This is probably incomplete, because I did this work on a plane and
don't have the 460 downloaded. We probably need to include schedule D as
well.