Always serialize OaklandCandidate with contribution calculations
By adding a `supporting_contribution_data` and
`opposing_contribution_data` to the candidate object's serialization, we
will make sure that we have top-line contribution numbers for any screen
we could need it. Plus it's not good API design to return two different
versions of the same object depending on the endpoint requested.

Fixes #17

cc @adborden @jnmarcus @kylew because of the backwards incompatibility