Add Form 497 to ReferendumSupportersCalculator
This is a somewhat complicated implementation, but out of necessity:
Form 497 rows do not contain the "Sup_Opp_Cd" column needed to bucket an
expenditure into either supporting or opposing a ballot measure. So we
must guess.

Also, I suspect this will double-count in some cases like Just Cause --
they gave money to a committee that is basically themselves. I don't
think we should count this money twice. I opened #29 to track this.