Update contributions_received stub to zero
Since there are no current records where "Sup_Opp_Cd" = 'O' for a
candidate in Schedule D or E, I'm going to just change this stub to
zero, which should get us through the v1 milestone.

We also won't see any records in the 496/497 that can be easily
attributed to be against a candidate because those files don't have the
"Sup_Opp_Cd" column.