Add simple page at / for health-checking reasons
I'm addressing some infrastructure debt on my personal server, and I'd
like to add a health check of this app. But it feels weird to have the
health-check alerting based on the /ballot/1 path (or anything
application-specific).

I also just accidentally thought the site was down by going to / and
seeing nothing. Turned out it was just a 404. To prevent accidents like
this in the future, I'm making the root URL return something arbitrary.