Remove ssconvert dependency
This replaces our ssconvert dependency with a tiny Ruby script that does
the same thing, except much slower. The performance implications of
parsing XLSX files (even a tiny 4 MB one like this) in Ruby are
unfortunate -- many of the Roo gem's methods just hung indefinitely on
my laptop as they allocated tons of memory and consumed a CPU.

But I think the script in this commit should work one-row-at-a-time and
so be a lot lower overhead. It takes about 3 minutes to convert the XLSX
file, but since it's only done once, it's not the worst thing in the
world.