Separate Makefile steps into a `download` chunk as well
In preparation for adding 2017 data as well, I'm simplifying further the
interpolation present in the Makefile since it is really hard to do
basic string manipulation in there.

I also want to ditch the `inputs` directory, since we can just consider
it a step in the downloading process to convert the XLSX file to CSVs.