Add check for ballot measure existence
For contributions to ballot measures that don't exist on the
spreadsheets, we won't be able to associate the records with a ballot
measure object. This is the case when we switch between years and there
are two sets of ballot measures in there... we'll need to come up with a
way to be able to look for the same ballot measure in different election
years. For now, let's just skip unknown ballot measures.