Download and process both 2016 and 2018 spreadsheets
This commit adds an "election_name" to the CSV sheets downloaded from
Google Sheets - the election name for 2016 is "oakland-2016" and for
2018 it is "oakland-2018".

csvkit once again impresses, by having the functionality to combine CSV
files and append a grouping column. This is how we'll add the
election_name column.

After that, it's just a matter of iterating over the two supported
elections and outputting the ballot JSON for each of them. Of course,
then we will have to figure out how to make the frontend use the right
one.