Create view "combined_contributions" to DRY up CTEs
The use of UNION ALL to combine the contributions resulted in some
changes to the build artifacts. I cross-checked with the original data
and it seemed like the additional results are correct -- in the case of
two donations of the same amount on the same day to the same committee,
we had previously uniq'd the second one out of existence with the
`UNION` but now we will report them both.