Add script to collect candidate information
This script takes one argument -- the election name -- and outputs the
CSV in the format of the current oakland_candidates.csv file.

Right now I have hardcoded a couple SF candidates into that file, but
ideally later we will be able to automatically extract the structure of
the ballot from the data.

This is a first step. The next are to 1) move logic for determining
Oakland candidates into the script as well, 2) start beefing up the
script with actual logic to infer the candidates, 3) test it by running
a build.