Update "Measure_Expenditures" view with 497 Support/Opp and recipients
This commit does two things:

1) Adds a "Recipient_Or_Description" column to the
"Measure_Expenditures" view, purely so we can give better debug
information during the processing run.

More notably, although the 497 data doesn't have the support/opposition
column, we can exploit the fact that it tends to have the Cmte_ID
(Committee ID) of the receiving committee. By adding the recipient
committee to the "committees" tab and annotating whether they support
the measure, we can use that Support / Opposition information. So:

2) Use the "Support or Oppose" column in the "committees" tab to
determine whether expenditures to that committee should be considered
to support or oppose the measure.