Add test suite and exclude in-kind IE's to candidates
There's a lot going on here, but now that I'm in the thick of it there's
no way to tease it apart.

The inciting change is to Fix #126 - a simple exclusion of
  "Expn_Code" = 'IKD' (In-Kind)
and
  "Expn_Code" = 'MON' (Monetary)

independent expenditures to candidates. Both are reported by the
candidate as a contribution, so we don't need to characterize them as
IE's in order to be counted.

And then to test it: this adds a test framework that essentially runs
`make import` and `make process` by shelling out before each test. The
framework imports a "fixture" directory into a separate
"disclosure-backend-test" database, the schema for which is copied from
"disclosure-backend". This means that the fixture directory doesn't need
to contain all the CSV files, only the ones with the data required for
the test and the metadata (oakland_candidates.csv).

I made a script to export a test case from the real data. It only does
Schedule D and needs more work.