Add wget wrapper that adds retry functionality
When we get a server error (e.g. 403 Forbidden as given by Google Docs
randomly), we'd like to retry. Modern versions of `wget` have a
`--retry-on-http-error` function that would do exactly what we need,
except that on Travis, this feature is not available in the distributed
version of wget.

This adds a gnarly, but likely effective, wrapper that tacks on retry
logic to wget.