Fix amendment deletion to handle null Filer_IDs
There was an unexpected condition where Cat Brooks amended a return and
also another candidate with a null Filer_ID filed a return with the same
"From_Date" and "Report_Num".

Since the semantics of NOT IN match present columns to null columns, the
previous Cat Brooks filing was not getting deleted.

This adds a test case for the record.

Fixes #138.