Add test case for Committee measure expenditures
This test is a regression test -- it is capturing the current
calculation logic that totals $136k for the committee, even though it
should be $145k. (A subsequent commit will update the test and its
implementation).

With this test, I clarified the process that the `oakland_*` objects
should be created in the Rspec test file, whereas all of the campaign
finance data should be dumped via the bin/export-test-data script.