Sort candidates by total_contributions, then by name
To improve the office_election list display, let's sort the candidates
in order by total_contributions. This is borne out by research
indicating that users are interested in the candidates that have raised
the most money locally. The remaining candidates are sorted by name so
they are in an expected order for a user searching to find a specific
one.