Post a comment with the build diff to Github
This will remove the need to run `make process` on your local branch -
instead Travis will run it and post it in a comment to the Pull Request.

The reason here is that there are some pesky build differences between
MacOS and Linux that have been very difficult to get rid of, so by
making the builds consistently running on Travis we can minimize
spurious diffs.