Extend the length of office_election title
We automatically create some rows in office_election during the
processing step, which isn't included in the inference used to create
the initial column lengths.

As such, the 31-character longest office in the data is too long for the
inferred 27-character max.

This, along with removal of non-integer FPPC numbers in the spreadsheet,
should fix the build.

[Fixes #187]