WIP Always include BMC expenditures
I'm not sure about this method. There are a lot of BMC (ballot measure
committees) that don't support ballot measures. (?) Curious what the
diff will look like on this one.

When calculating referendum supporters (using the Measure_Expenditures
view), we included:

1. E-Expenditures from committees who reported a "Bal_Name" that was on
the "name_to_number" sheet.
2. 496 independent expenditures as determined by the "Bal_Name" matching
the "name_to_number" sheet.
3. E-Expenditures from committees using "name_to_number" to match when
"Bal_Name" is null.

...but we seemingly do not include expenditures from ballot measure
committees. We link it up in the spreadsheet, so I think we can just
assume that any expenditures from BMC's are for the purpose of their
associated ballot measure.