Add ability for calculator to specify dependencies
This is a light implementation to allow calculators to define
dependencies on other calculators, for example to allow us to use
calculations from one object (e.g. candidates) in the calculations for
another object (e.g. an election).

To specify dependencies, calculators can define a `dependencies` class
method in the format in the ElectionTotals calculator, like so:

```ruby
def self.dependencies
  [
    { model: Candidate, calculation: :total_contributions }
  ]
end
```

This calculator will then not be run until another calculator saves a
calculation with that name for a candidate. (This operates under the
assumption that every calculation is created by only a single
calculator.)

If a calculator cannot be run due to unmet dependencies, a helpful error
message will be thrown.