Output contribution list for Candidates without committee rows
For candidates without entries in the `committees` spreadsheet tab, we
still want to output their contribution list. The problem is, we store
their contribution list on the `Candidate` instead of their associated
committee. We also muddy the waters by saving a calculation called
`total_contributions` which is calculated differently between Candidates
and Committees.

This commit:
* Renames the calculation for the total of the contribution list to
  `contribution_list_total` and adds fallback logic to mimic the
  previous implicit overwriting behavior
* Adds a `committee_data` method to Candidate to output the calculations
  related to the committee.